refactor: rename auth-wall dir to auth because it contains a hook too
