+++ src/app/app.js
@@ -2,7 +2,7 @@ import { CssVariables } from '@dhis2/ui'
 import React from 'react'
 import { QueryParamProvider } from 'use-query-params'
 import { ActionBar } from '../action-bar'
-import { AuthWall } from '../auth-wall'
+import { AuthWall } from '../auth'
 import { ContextSelector } from '../context-selector'
 import { CurrentUserProvider } from '../current-user'
 import { DataWorkspace } from '../data-workspace'
